Remove "focus_gained" action
This has been commented out since we took over development of the extension,
so I don't really see the harm in removing it completely - especially as we can
accomplish the same thing using the html5 page visibility api. I also removed the
visibilityChange function - we can check whether document.hidden is true in
the clearUnreadPosts function.