Add dramalinks fixes
Previously, it was possible for the dramalinks function to
respond before the XHR had completed - this would lead to
"undefined" being displayed in place of the ticker (until the
page was refreshed, returning the cached dramalinks value).

To fix this, we set up a listener before we request "dramalinks"
from background page - once the "updatedrama" message
has been returned (signifying that the XHR has completed)
we request the cached dramalinks using updateDramaTicker.