Improve imagemap searching, add options menu
Added a "filename" property which allows for more accurate searching (as it
contains the original fullsize file extension & removes URL encoding). It would
also be possible to allow user to edit this without affecting the caching function.

Options menu currently displays table containing list of filenames and URLs,
size on disk, and a button to empty the cache.