Improve performance of Browse Imagemap feature
The sendToEncoder method will only send images to the base64
encoder if they haven't been cached before (by checking for the oldsrc attribute).

It also saves encoded images as they are encoded, instead of waiting for
the sendToEncoder loop to finish. This makes the code cleaner as it requires
fewer parameters to be passed between methods.