Split synced config object if it exceeds QUOTA_BYTES_PER_ITEM
One problem with current implementation is that a key-value pair that exceeds
QUOTA_BYTES_PER_ITEM would throw an error - in this situation we would need
to split the key-value pair into parts. While it seems unlikely that this
would ever happen, we should still probably prepare for it.

Needs refactoring.