Use CSS for user highlights/etc, change highlight behaviour
The CSS makes the process of highlighting much simpler - we just need to check that
the value we're looking for is contained by the element, and then set the appropriate
attribute (instead of having to iterate over multiple sets of HTMLCollections and
set style properties to style each table row).

Keyword/tag highlights are now applied to their corresponding anchor tags as well -
this means that we can allow for multiple types of highlight to be displayed in same
table row (instead of having to 'override' each type of highlight with the next). Still
needs tweaking.