Rewrite database methods to use separate object stores for searching and caching
We create a database which only contains the filenames and src key for each image -
then, if we need to retrieve the full image data, we can just get the records directly
(instead of iterating over the full database to find what we need). Search performance is
roughly 5x faster using this method.